Less zealous HTTP caching when memory cache miss
If the cache was just cleared (through refresh or reschedule button) or
the server restarted, then we should not assume that the client's last
modified actually holds complete truth over what should be shown.
Especially when the schedule gets changed (either by a reschedule or by
an agent being completely finished), then we should refresh the entire
page (at the cost of expensive collection).